requirements.txt
python-telegram-bot==20.8
SQLAlchemy==2.0.27
pydantic==2.6.1
python-dotenv==1.0.1
pytest==8.0.2
black==24.2.0
isort==5.13.2
mypy==1.8.0
Faker==22.6.0 